+++ src/StlSolid.cpp
@@ -357,7 +357,7 @@ HeeksObj* CStlSolid::ReadFromXMLElement(TiXmlElement* pElem)
 	for(TiXmlAttribute* a = pElem->FirstAttribute(); a; a = a->Next())
 	{
 		std::string name(a->Name());
-		if(name == "col"){c = HeeksColor(a->IntValue());}
+		if(name == "col"){c = HeeksColor((long)(a->IntValue()));}
 	}
 
 	CStlSolid* new_object = new CStlSolid(&c);
